+++ src/library/vm/vm_string.h
@@ -11,4 +11,7 @@ Author: Leonardo de Moura
 namespace lean {
 std::string to_string(vm_obj const & o);
 vm_obj to_obj(std::string const & str);
+
+void initialize_vm_string();
+void finalize_vm_string();
 }